formula: Stub support for TimeVariables
If a variable is read with ModelVariable::read(), we check if it
matches a file name first, if so read the timevariable. Otherwise fall
back to reading formula. Mine through the model path so checking for
this is possible.

This will end up supporting TimeVariables more often than gadget2 does,
but that's unlikely to be a problem.